Add support for DNS name restriction to CA gen.
This lets you issue a CA cert that restricts itself to a particular domain.
For instance, a cert can be issued only good for *.pao1.playground.global,
making it reasonably safe to import into Chrome as a trusted authority,
without the risk of a MITM attack on arbitrary web sites.